+++ ui/main/game/galactic_war/shared/js/gw_faction_0.js
@@ -25,40 +25,28 @@ define([
     {
       name: "Able",
       character: "!LOC:Armor",
-      color: [
-        [204, 255, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.armour,
       commander: "/pa/units/commanders/imperial_able/imperial_able.json",
     },
     {
       name: "AceAI",
       character: "!LOC:Roboticist",
-      color: [
-        [153, 255, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.roboticist,
       commander: "/pa/units/commanders/imperial_aceal/imperial_aceal.json",
     },
     {
       name: "Alpha",
       character: "!LOC:Uber",
-      color: [
-        [102, 255, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.uber,
       commander: "/pa/units/commanders/imperial_alpha/imperial_alpha.json",
     },
     {
       name: "Chronoblip",
       character: "!LOC:Fabber",
-      color: [
-        [0, 204, 204],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.fabber,
       commander:
         "/pa/units/commanders/imperial_chronoblip/imperial_chronoblip.json",
@@ -66,10 +54,7 @@ define([
     {
       name: "Mjon",
       character: "!LOC:Defender",
-      color: [
-        [0, 153, 153],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       econ_rate: 1,
       personality: personalities.defender,
       commander: "/pa/units/commanders/imperial_mjon/imperial_mjon.json",
@@ -77,20 +62,14 @@ define([
     {
       name: "Delta",
       character: "!LOC:Luddite",
-      color: [
-        [153, 204, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.luddite,
       commander: "/pa/units/commanders/imperial_delta/imperial_delta.json",
     },
     {
       name: "Enzomatrix",
       character: "!LOC:Technologist",
-      color: [
-        [102, 178, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.technologist,
       commander:
         "/pa/units/commanders/imperial_enzomatrix/imperial_enzomatrix.json",
@@ -98,10 +77,7 @@ define([
     {
       name: "Fiveleafclover",
       character: "!LOC:Cautious",
-      color: [
-        [51, 153, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.cautious,
       commander:
         "/pa/units/commanders/imperial_fiveleafclover/imperial_fiveleafclover.json",
@@ -109,50 +85,35 @@ define([
     {
       name: "Gamma",
       character: "!LOC:Aggressive",
-      color: [
-        [0, 128, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.aggressive,
       commander: "/pa/units/commanders/imperial_gamma/imperial_gamma.json",
     },
     {
       name: "Gnugfur",
       character: "!LOC:Rush",
-      color: [
-        [0, 102, 204],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.rush,
       commander: "/pa/units/commanders/imperial_gnugfur/imperial_gnugfur.json",
     },
     {
       name: "Nemicus",
       character: "!LOC:Turtle",
-      color: [
-        [0, 76, 153],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.turtle,
       commander: "/pa/units/commanders/raptor_nemicus/raptor_nemicus.json",
     },
     {
       name: "Kapowaz",
       character: "!LOC:Original",
-      color: [
-        [0, 0, 153],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.legonisMachinaOriginal,
       commander: "/pa/units/commanders/imperial_kapowaz/imperial_kapowaz.json",
     },
     {
       name: "JT100010117",
       character: "!LOC:Absurd",
-      color: [
-        [0, 0, 204],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.absurd,
       commander:
         "/pa/units/commanders/imperial_jt100010117/imperial_jt100010117.json",
@@ -160,10 +121,7 @@ define([
     {
       name: "Kevin4001",
       character: "!LOC:Factory",
-      color: [
-        [0, 0, 225],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.factory,
       commander:
         "/pa/units/commanders/imperial_kevin4001/imperial_kevin4001.json",
@@ -171,10 +129,7 @@ define([
     {
       name: "Mostlikely",
       character: "!LOC:Swarm",
-      color: [
-        [51, 51, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.swarm,
       commander:
         "/pa/units/commanders/imperial_mostlikely/imperial_mostlikely.json",
@@ -182,10 +137,7 @@ define([
     {
       name: "Nagasher",
       character: "!LOC:Economist",
-      color: [
-        [204, 229, 255],
-        [192, 192, 192],
-      ],
+      color: factionColour,
       personality: personalities.economist,
       commander:
         "/pa/units/commanders/imperial_nagasher/imperial_nagasher.json",
@@ -196,10 +148,7 @@ define([
   const randomAI = {
     name: "Aryst0krat",
     character: "!LOC:Random",
-    color: [
-      [0, 255, 255],
-      [192, 192, 192],
-    ],
+    color: factionColour,
     personality: randomMinion.personality,
     commander:
       "/pa/units/commanders/imperial_aryst0krat/imperial_aryst0krat.json",